cadastro_perguntas/requirements.txt
Django==3.2.*
djangorestframework
markdown
django-filter
django-guardian
Pygments
python-dotenv
gunicorn
autopep8
django-graphql-jwt==0.4.0
Pillow
django-cors-headers
pyyaml
uritemplate
xhtml2pdf
pytest
pytest-django
python-decouple
python-gettext
graphene-django==3.2.1
psycopg2-binary
django-allauth==0.50.0
django-crispy-forms==1.14.0
django-graphiql==0.4.4
django-widget-tweaks==1.4.12
